mysql生成修正 启动参数修正 html，php界面生成模板修改。

+++ ftl/PHPremove.ftl
@@ -1,5 +1,9 @@
 ﻿<?php
+$error_msg = '{"result" : "FAIL", "msg" : "$1"}';
+
 include_once '../../common/common_include.php';
+require_once '../../common/check_access.php';
+check_access('/${entity.dbschema}/${entity.name}');
 
 $sql = 'delete from ${entity.tableName} 
 where ${entity.idname} = :${entity.idname}';
@@ -13,8 +17,7 @@ try {
 	$result = common_commit($stmt);
 } catch(PDOExecption $e) { 
 	$db->rollback(); 
-	header('Status: 600');
-    echo '{"code" : 600, "msg" : $e->getMessage()}';
+    echo '{"code" : 600, "msg" : "$e->getMessage()"}';
     exit();
 }
